+++ spec/chainable_queries_spec.rb
@@ -129,6 +129,48 @@ describe ContentfulModel::ChainableQueries do
         expect(subject.find_by(foo: {gte: 123})).to eq subject
         expect(subject.query.parameters).to include('fields.foo[gte]' => 123)
       end
+
+      it 'when multiple fields' do
+        expect(subject.find_by(foo: true, bar: 123)).to eq subject
+        expect(subject.query.parameters).to include('fields.foo' => true , 'fields.bar' => 123)
+      end
+
+      it 'supports sys fields' do
+        expect(subject.find_by('sys.id': 'foo')).to eq subject
+        expect(subject.query.parameters).to include('sys.id' => 'foo')
+      end
+    end
+
+    describe '::paginate' do
+      it 'defaults to first page and 100 items and sort by updatedAt' do
+        expect(subject.paginate).to eq subject
+        expect(subject.query.parameters).to include('limit' => 100, 'skip' => 0, 'order' => 'sys.updatedAt')
+      end
+
+      it 'requesting second page will add page_size to skip' do
+        expect(subject.paginate(2)).to eq subject
+        expect(subject.query.parameters).to include('limit' => 100, 'skip' => 100)
+      end
+
+      it 'can change page_size' do
+        expect(subject.paginate(1, 20)).to eq subject
+        expect(subject.query.parameters).to include('limit' => 20, 'skip' => 0)
+      end
+
+      it 'can change page_size and select a different page' do
+        expect(subject.paginate(3, 20)).to eq subject
+        expect(subject.query.parameters).to include('limit' => 20, 'skip' => 40)
+      end
+
+      it 'outliers are changed to default values' do
+        expect(subject.paginate(-1, 'foo')).to eq subject
+        expect(subject.query.parameters).to include('limit' => 100, 'skip' => 0)
+      end
+
+      it 'can sort by a different field' do
+        expect(subject.paginate(1, 100, 'sys.createdAt')).to eq subject
+        expect(subject.query.parameters).to include('limit' => 100, 'skip' => 0, 'order' => 'sys.createdAt')
+      end
     end
 
     describe '::search' do